subch

subch

Usage

subch(g, fig_height = 7, fig_width = 5)

Arguments

g

figure to be included as a subchunk in rmarkdown

fig_height

height of the figure

fig_width

width of the figure

Value

a figure to be included as a subchunk in rmarkdown