add_iycf

add_iycf

Usage

add_iycf(
  .dataset,
  age_months = "age_months",
  iycf_1 = "iycf_1",
  iycf_2 = "iycf_2",
  iycf_3 = "iycf_3",
  iycf_4 = "iycf_4",
  iycf_5 = "iycf_5",
  iycf_6a = "iycf_6a",
  iycf_6b = "iycf_6b",
  iycf_6c = "iycf_6c",
  iycf_6d = "iycf_6d",
  iycf_6e = "iycf_6e",
  iycf_6f = "iycf_6f",
  iycf_6g = "iycf_6g",
  iycf_6h = "iycf_6h",
  iycf_6i = "iycf_6i",
  iycf_6j = "iycf_6j",
  iycf_7a = "iycf_7a",
  iycf_7b = "iycf_7b",
  iycf_7c = "iycf_7c",
  iycf_7d = "iycf_7d",
  iycf_7e = "iycf_7e",
  iycf_7f = "iycf_7f",
  iycf_7g = "iycf_7g",
  iycf_7h = "iycf_7h",
  iycf_7i = "iycf_7i",
  iycf_7j = "iycf_7j",
  iycf_7k = "iycf_7k",
  iycf_7l = "iycf_7l",
  iycf_7m = "iycf_7m",
  iycf_7n = "iycf_7n",
  iycf_7o = "iycf_7o",
  iycf_7p = "iycf_7p",
  iycf_7q = "iycf_7q",
  iycf_7r = "iycf_7r",
  iycf_8 = "iycf_8",
  iycf_6c_swt = "iycf_6c_swt",
  iycf_6d_swt = "iycf_6d_swt",
  iycf_6h_swt = "iycf_6h_swt",
  iycf_6j_swt = "iycf_6j_swt",
  yes_value = "yes",
  no_value = "no",
  dnk_value = "dont_know",
  pna_value = "pna",
  iycf2_immediate_value = "immediately",
  iycf2_lessday_value = "less_than_one_day",
  iycf2_moreday_value = "more_than_one_day",
  grouping = NULL,
  uuid = "uuid",
  loop_index = NULL
)

Arguments

.dataset

the raw child nutrition loop dataset with all iycf indicators

age_months

the name of the variable that indicates the age of child per month. By default it is age_months

iycf_1

the name of the variable that indicates if the child ever breastfed By default "iycf_1"

iycf_2

the name of the variable that indicates how long the child started breastfeeding after birth. By default "iycf_2"

iycf_3

the name of the variable that indicates Exclusive Breastfeeding First 2 Days By default "iycf_3"

iycf_4

the name of the variable that indicates if the child was breastfed yesterday By default "iycf_4"

iycf_5

the name of the variable that indicates if the child drink anything from a bottle yesterday. By default "iycf_5"

iycf_6a

the name of the variable that indicates if the child had water yesterday. By default "iycf_6a"

iycf_6b

the name of the variable that indicates if the child had infant formula yesterday. By default "iycf_6b"

iycf_6c

the name of the variable that indicates if the child had milk from animal yesterday. By default "iycf_6c"

iycf_6d

the name of the variable that indicates if the child had yogurt drink yesterday. By default "iycf_6d"

iycf_6e

the name of the variable that indicates if the child had chocolate flavored drink yesterday. By default "iycf_6e"

iycf_6f

the name of the variable that indicates if the child had fruit juice or fruit flavored drink yesterday. By default "iycf_6f"

iycf_6g

the name of the variable that indicates if the child had sodas/malt/sports/energy drink yesterday. By default "iycf_6g"

iycf_6h

the name of the variable that indicates if the child had tea/coffee/herbal drink yesterday. By default "iycf_6h"

iycf_6i

the name of the variable that indicates if the child had clear broth/soup yesterday. By default "iycf_6i"

iycf_6j

the name of the variable that indicates if the child had other liquids yesterday. By default "iycf_6i"

iycf_7a

the name of the variable that indicates if the child had yogurt as food yesterday. By default "iycf_7a"

iycf_7b

the name of the variable that indicates if the child had porridge/bread/rice/noodles/pasta yesterday. By default "iycf_7b"

iycf_7c

the name of the variable that indicates if the child had pumpkin/carrots/sweet red peppers/squash/sweet potato yesterday. By default "iycf_7c"

iycf_7d

the name of the variable that indicates if the child had plantains/white potato/yams/manioc/cassava yesterday. By default "iycf_7d"

iycf_7e

the name of the variable that indicates if the child had dark green leafy vegetables yesterday. By default "iycf_7e"

iycf_7f

the name of the variable that indicates if the child had other vegetables yesterday. By default "iycf_7f"

iycf_7g

the name of the variable that indicates if the child had ripe mangoes/ripe papayas yesterday. By default "iycf_7g"

iycf_7h

the name of the variable that indicates if the child had other fruits yesterday. By default "iycf_7h"

iycf_7i

the name of the variable that indicates if the child had liver/kidney/heart yesterday. By default "iycf_7i"

iycf_7j

the name of the variable that indicates if the child had sausage/hot dogs/ham/bacon/salami/canned meat yesterday. By default "iycf_7j"

iycf_7k

the name of the variable that indicates if the child had other meat/beef/pork/lamb/goat/chicken/duck yesterday. By default "iycf_7k"

iycf_7l

the name of the variable that indicates if the child had eggs yesterday. By default "iycf_7l"

iycf_7m

the name of the variable that indicates if the child had fresh/dried/shell fish yesterday. By default "iycf_7m"

iycf_7n

the name of the variable that indicates if the child had beans/peas/lentils/nuts/seeds yesterday. By default "iycf_7n"

iycf_7o

the name of the variable that indicates if the child had hard/soft cheese yesterday. By default "iycf_7o"

iycf_7p

the name of the variable that indicates if the child had sweet foods yesterday. By default "iycf_7p"

iycf_7q

the name of the variable that indicates if the child had chips/crisps/puffs/french fries yesterday. By default "iycf_7q"

iycf_7r

the name of the variable that indicates if the child had other solid food yesterday. By default "iycf_7q"

iycf_8

the name of the variable that indicates if the meal frequency the child had yesterday. By default "iycf_8"

iycf_6c_swt

the name of the variable that indicates if the child had sweet milk yesterday. By default "iycf_6c_swt"

iycf_6d_swt

the name of the variable that indicates if the child had sweet yogurt drink yesterday. By default "iycf_6d_swt"

iycf_6h_swt

the name of the variable that indicates if the child had sweet tea drink yesterday. By default "iycf_6h_swt"

iycf_6j_swt

the name of the variable that indicates if the child had other sweet drink yesterday. By default "iycf_6j_swt"

yes_value

the value of the choice "yes" to all the liquid/food categories

no_value

the value of the choice "no" to all the liquid/food categories

dnk_value

the value of the choice "dont know" to all the liquid/food categories

pna_value

the value of the choice "prefer not to answer" to all the liquid/food categories

iycf2_immediate_value

the value of the choice "immediately" to the indicator iycf_2

iycf2_lessday_value

the value of the choice "less than a day" to the indicator iycf_2

iycf2_moreday_value

the value of the choice "more than a day"" to the indicator iycf_2

grouping

the name of the variable that indicates the grouping variable - usually "enumerator"

uuid

uuid variable

loop_index

unique identifier for each individual in the iycf loop. By default, loop_index

Value

a dataframe that includes all the added iycf indicators. This includes: - iycf_evbf / Ever Breastfed - iycf_eibf / Early Initiation of Breastfeeding - iycf_ebf2d / Exclusive Breastfeeding First 2 Days After Birth - iycf_ebf / Exclusive Breastfeeding - iycf_mixmf / Mixed Milk Feeding - iycf_cbf / Continued Breastfeeding 12-23 months - iycf_isssf / Introduction of Solid, Semi-Solid, or Soft Foods - iycf_mdd_score / Minimum Dietary Diversity 6-23 months Score - iycf_mdd_cat / Minimum Dietary Diversity 6-23 months Category - iycf_mmf / Minimum Meal Frequency 6-23 months - iycf_mmff / Minimum Milk Feeding Frequency For Non-Breastfed Children 6-23 months - iycf_mad / Minimum Acceptable Diet 6-23 months - iycf_eff / Eggs & Flesh Foods Consumption 6-23 months - iycf_swb / Sweet Beverage Consumption 6-23 months - iycf_ufc / Unhealthy Food Consumption - iycf_zvf / Zero Vegetable or Fruit Consumption 6-23 months - iycf_bof / Bottle Feeding 0-23 months

Examples

if (FALSE) add_iycf(df1)